#14 로그인 시 userId 반환 기능 구현

+++ src/main/java/com/sayup/SayUp/controller/FriendshipController.java
@@ -22,7 +22,7 @@ public class FriendshipController {
     private final FriendshipService friendshipService;
 
     @PostMapping("/request/{addresseeId}")
-    public ResponseEntity<?> sendFriendRequest(@AuthenticationPrincipal UserDetails userDetails, @PathVariable Integer addresseeId) {
+    public ResponseEntity<?> sendFriendRequest(@AuthenticationPrincipal UserDetails userDetails, @PathVariable Long addresseeId) {
         if (userDetails == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("User not authenticated");
         }
